Move message task to tasks/messages
